Queuebert filter fixes.
Added better high level testing around search filters and fixed the
problem of not having a way to pass regular expression filters in the
JSON messages used by Queuebert. Also clarified the fact that string
filters are supposed to match any part of a path and not only the
entire path.